Passed all units test

+++ src/lifeintruderservice/repositories/person.py
@@ -14,7 +14,7 @@ class PersonRepository:
 
     @staticmethod
     def get(guid):
-        """ Query a person by the globally unique identifier (GUID)"""
+        """ Query a person by the globally unique identifier (GUID) """
 
         logger.info('Retreiving indivual informations:"{0}"'.format(guid))
 
@@ -25,8 +25,8 @@ class PersonRepository:
         except RqlRuntimeError as e:
             abort(404, f'Records for globally unique identifier (GUID) "{guid}" were not found. "{e}"')
 
-    @staticmethod
-    def create(guid, profile):
+
+    def create(self, guid, profile):
         """ Create a new user """
 
         try:
@@ -39,9 +39,9 @@ class PersonRepository:
         except RqlRuntimeError as e:
             abort(503, f'Record could not be inserted. Message: {e}')
 
-    @staticmethod
+
     def update(self, guid, profile):
-        """ Update a person's profile"""
+        """ Update a person's profile """
         
         logger.info(f'Updating individual information:"{guid}"')
         
@@ -53,15 +53,19 @@ class PersonRepository:
             cursor = rdb.table(args.rethinkdb_table).filter({args.rethinkdb_primary_key: guid}).update(person).run(g.rdb_conn)
             return cursor
         except RqlRuntimeError as e:
-            abort(503, f'Record could not be update. Message: {e}')
-    
-    @staticmethod
-    def remove(guid):
-        """ Remove a person by its globally unique identifier (GUID)"""
+            abort(503, f'Record could not be updated. Message: {e}')
 
-        logger.info('Remove all individual related:"{0}"'.format(guid))
 
+    def remove(self, guid):
+        """ Remove a person by its globally unique identifier (GUID) """
+        
+        logger.info(f'Removing individual information:"{guid}"')
+        
+        if self.get(guid) == None:
+            abort(404, f'No existing record existing record has been found for GUID {guid}.')
+        
         try:
-            rdb.table(args.rethinkdb_table).filter({args.rethinkdb_primary_key: guid}).delete().run(g.rdb_conn)
+            cursor = rdb.table(args.rethinkdb_table).filter({args.rethinkdb_primary_key: guid}).delete().run(g.rdb_conn)
+            return cursor
         except RqlRuntimeError as e:
-            abort(404, f'Records for globally unique identifier (GUID) "{guid}" were not found. "{e}"')
+            abort(503, f'Record could not be removed. Message: {e}')
